+++ src/main/java/baseball/util/Constants.java
@@ -1,4 +1,4 @@
-package baseball.utils;
+package baseball.util;
 
 public final class Constants {
     /*
